requirements.txt
confluent-kafka>=2.3.0
playwright>=1.30.0
pydantic>=2.0.0
numpy>=1.24.0
opencv-python
pillow
torch
torchvision
transformers